+++ src/view/point-edit/templates/create-destination-template.js
