+++ src/view/point-edit/templates/create-offers-template.js
@@ -1,0 +1,37 @@
+export const createOffersTemplate = (offers) => {
+  return (
+    `<section class="event__section  event__section--offers">
+      <h3 class="event__section-title  event__section-title--offers">
+        Offers
+      </h3>
+      <div class="event__available-offers">
+        ${offers
+          .map((offer) => {
+            const {key, name, price} = offer;
+
+            return (
+              `<div class="event__offer-selector">
+                <input
+                  class="event__offer-checkbox visually-hidden"
+                  id="event-offer-${key}-1"
+                  value="${key}"
+                  type="checkbox" name="event-offer-${key}"
+                  ${offer.isActivated ? `checked` : ``}
+                >
+                 <label class="event__offer-label" for="event-offer-${key}-1">
+                  <span class="event__offer-title">
+                    ${name}
+                  </span>
+                  &plus;
+                  &euro;&nbsp;<span class="event__offer-price">
+                    ${price}
+                 </span>
+                </label>
+              </div>`
+            );
+          })
+          .join(``)}
+      </div>
+    </section>`
+  );
+};
