+++ src/view/point-edit/templates/create-reset-button-template.js
@@ -1,0 +1,7 @@
+export const createResetButtonTemplate = (isAddMode) => {
+  return (
+    `<button class="event__reset-btn" type="reset">
+      ${isAddMode ? `Cancel` : `Delete`}
+    </button>`
+  );
+};
