+++ src/view/point-edit/templates/create-rollup-button-template.js
@@ -1,0 +1,9 @@
+export const createRollupButtonTemplate = () => {
+  return (
+    `<button class="event__rollup-btn" type="button">
+      <span class="visually-hidden">
+      Open event
+      </span>
+    </button>`
+  );
+};
