+++ src/view/point-edit/templates/create-save-button-template.js
@@ -1,0 +1,11 @@
+export const createSaveButtonTemplate = (isDisabled) => {
+  return (
+    `<button
+      class="event__save-btn  btn  btn--blue"
+      type="submit"
+      ${isDisabled ? `disabled` : ``}
+    >
+      Save
+    </button>`
+  );
+};
