+++ src/view/point/templates/create-offer-template.js
